fixed issue #295 Color Scheme Switching does not automatically refresh and may lead to an infinity loop
FeatureIDE was not responding when changing a color scheme and saving changes to the model. This was caused by call to Thread.stop in AStoppableJob class which leads to damaged objects (ThreadDeath exception occurred) and uncontrolled behavior, which manifest into some kind of race condition. However, changing stop to join solves the problem. Although it now can happen, that a stoppable job isn't stopped if it was not able to join the inner thread.