FeatureIDE might be evolve to be feature-oriented composable. The first step is to automatically generate a feature model xml file which contains file-dependencies.
In long-run this model can be extended to a "client-oriented model" containing the feature "Modeling Tools" for instance. These "client-oriented" features will be constrain the feature model file-dependency model. A composer (script,..) can read a valid project configuration, and copy required files to the product bundle.

This commit contains an initial prototypic scanner which scans the FeatureIDE for plugins. Each plugin might contain: (1) asset files, (2) a manifest file for the eclipse plugin system, (3) source files, and (4) external dependency. The scanner builds a set of statements on these entities. More in detail: The scanner generates a feature called "featureide". This feature is the root for (a) external dependencies feature, and (b) a set of plugin features (sub projects). Each element in (b) contains of a feature indicating the manifest file, the asset, and the the source. Asset files are directly attached to the asset feature (i.e., images are assets and required). Source files are organized as follows: Each source file is a feature with the full qualified class name, which has as its parent the package (which is also a feature as child of the source feature). Each source file requires a set of other classes which form a dependency to other features. These features are either known (as part of feature ide), or can not be fully resolved (as part of external dependency). Hence, a source file requires (implies) other source files either in a plugin of feature ide, or in the external dependency feature.