Preparations to migrate LOC nodes to new datatype FileFeatureLOCMapper.
Added more functions to datatype FileFeatureLOCMapper and rebuild some
of its structure because of changed requirements.

I also began to implement the initChildren() function in the new
LOCFilterNode and LOCFilterChildNode, which begin to replace the old
HashMapNodeTwoStrings classes.

There are a few questions left though if the FileFeatureLocMapper is
filled with data correctly. Tests are required.