Clear explanations before refreshing after changes critical to analysis
When a feature model element figure is being refreshed, it may print an
explanation. However, if the figure is being refreshed because of a
change in the feature diagram that affects analysis, the explanation
might no longer accurately reflect the defect in the feature diagram.
This is particularly troublesome if the explanation references a feature
model element that has just been deleted and thus cannot be resolved
when printing the explanation. To solve this, by clearing all
explanations, they are not printed until after the feature diagram
refresh and their recreation during the analysis following it.

Alternatively, the explanations could be updated before being printed by
executing analysis before as opposed to after the feature diagram
refresh. However, changing the established order might entail unforeseen
side effects.