Explain using clause traces instead of literal annotations
Previously, the information for generating explanations (such as which
part of the formula originate from which cross-tree constraint) were
stored inside the literal. This was implemented as an integer storing
the constraint index or the role of the source feature in a child
relationship.

However, storing such information in the literal object fails to
separate the concerns. A literal should not require knowledge of its
origin. Additionally, a literal could come not only from the feature
model but other applications as well, which means each application
would have to annotate the literal for its case-specific information. It
also increases the footprint of the literals, which are used throughout
the entire application.

Another dimension to the problem is the fact that the granularity of
literals is too fine. For instance, the same information was stored for
each literal of the same constraint in the entire target formula. It
would be more convenient and space-efficient to store this information
for each clause instead of each literal.

As such, the literal annotations are replaced by clause traces. To this
end, a trace model is implemented. During transformation from feature
model to formula, it keeps track of the source element to target clause
mapping. This information is then used in place of the literal
annotations when generating explanations.

As a result, the literals do not require knowledge of how
transformations work. Also, applications do not require knowledge of how
literal annotations work; instead, they use the more abstract and
generalizable trace model.