Remember clause indexes instead of padding with null
Because Sat4J does not free up a clause's index upon removal from the
solver, the SAT solver facade needs to remember the already used indexes
even after removing the corresponding clauses. Previously, this was done
by setting the clause in the internal clause list to null. However, this
requires filtering out the null values when working with the clauses.
This padding increases the more clauses are removed and as such causes a
performance degradation the longer the instance of the SAT solver facade
is used.

Now, the null padding is removed. Instead, the mapping between the Sat4J
clause indexes and the internal clause indexes is stored in two maps.
This fixes the performance degradation. For sparse clause indexes, i.e.,
after many clauses have been removed, this also decreases the memory
footprint because neither the null padding nor the removed map entries
are stored.

Additionally, this fixes a mismatch between the internally stored
clauses and the constraints returned by the Sat4J solver when a clause
is added to the SAT solver facade that cannot be added to the Sat4J
solver as well (because it is a duplicate, a contradiction, etc.). In
such a case, instead of attempting to model the index through the
internal clause list, the SAT solver facade now simply does not remember
any Sat4J clause indexes for that clause. This also makes it no longer
necessary to ignore duplicates added to the SAT solver facade.