Workaround for #1003
As the export functionality is broken under Linux due to a bug in the
GTK framework, this commit implements a workaround.
The workaround consists of:
* When exporting a feature model an info dialog appears (which can be
disabled by toggling) to inform the user of the workaround
* If the user is running FeatureIDE on a Linux machine, instead of the
SWT FileDialog, an instance of NonGTKFileDialog is instantiated, which
is wrapper for Swing's FileChooser.

Notably, the workaround is implemented cleanly and the Swing dialog is
encapsulated in its own, asynchronous thread.