Fixes #1178 (#1203)
* Fixes #1178

The configuration map now says "Please open a FeatureIDE project" when
opening a featuremodel from a different project in the feature diagram
editor.
Additionally, an issue with refreshing the view when closing all editors
is fixed.

* Fixed a bug where the text did not refresh wenn closing an editor with a
non featureIDE project

* Fixes #1178 and a few update problems of the configuration map

Now the string shown in the configuration map is always "Please open a
FeatureIDE project". Additionally, this fix involves updating issues
with the configuration map.