Fixes #1180 #1181 #1174 #1178 #1195 (#1219)
* Fixes #1181 (#1205/#1208)

When declaring the legend size, we set the width with regards to the
String "Alternative Group" not "Alternative".
Additionally, now the width is set by the word "Or Group" not "Optional"
for "Or Group". (Was a not recognized mistake since this size is
probably never declared)

Co-authored-by: Rahel <64740628+Rahel12342@users.noreply.github.com>

* Fixes #1181

* Fixes #1180

When declaring the legend size, we set the width with regards to the
String "Alternative Group" not "Alternative".
Additionally, now the width is set by the word "Or Group" not "Optional"
for "Or Group". (Was a not recognized mistake since this size is
probably never declared)

* Fixes issue #1180

Co-authored-by: Rahel <64740628+Rahel12342@users.noreply.github.com>
Co-authored-by: rahel <rahel.arens@web.de>

* Fixes #1174 (#1202) (#1216)

* Fixes #1174 (#1202)

Only allow direct editing for features that are not external features.

Co-authored-by: Rahel <64740628+Rahel12342@users.noreply.github.com>
Co-authored-by: rahel <rahel.arens@web.de>

* Fixes #1178 (#1203)  (#1217)

* Fixes #1178 (#1203)

The configuration map now says "Please open a FeatureIDE project" when
opening a featuremodel from a different project in the feature diagram
editor.
Additionally, an issue with refreshing the view when closing all editors
is fixed.

* Fixed a bug where the text did not refresh wenn closing an editor with a
non featureIDE project

* Fixes #1178 and a few update problems of the configuration map

Now the string shown in the configuration map is always "Please open a
FeatureIDE project". Additionally, this fix involves updating issues
with the configuration map.

Co-authored-by: Rahel <64740628+Rahel12342@users.noreply.github.com>
Co-authored-by: rahel <rahel.arens@web.de>

* Fixes #1195 (#1218)

* Fixes #1195

Select All, Delete and Undo/Redo are now redirected to the text cell (#1198)
editor when renaming a feature

Co-authored-by: Rahel <64740628+Rahel12342@users.noreply.github.com>
Co-authored-by: rahel <rahel.arens@web.de>
Co-authored-by: johannesherschel <84278649+johannesherschel@users.noreply.github.com>
Co-authored-by: Johannes Herschel <johannes.herschel@uni-ulm.de>